pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gelato"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.12.0",
    "bitsandbytes>=0.49.2",
    "cairosvg>=2.9.0",
    "datasets>=4.5.0",
    "levenshtein>=0.27.3",
    "music21>=9.9.1",
    "numpy>=2.4.1",
    "pillow>=12.1.0",
    "sentencepiece>=0.2.1",
    "tensorboard>=2.20.0",
    "timm>=1.0.25",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.48.3",
    "verovio==6.0.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.14",
    "ty>=0.0.13",
]

[tool.uv.sources]
torch = [{ index = "pytorch-cu130" }]
torchvision = [{ index = "pytorch-cu130" }]

[[tool.uv.index]]
name = "pytorch-cu130"
url = "https://download.pytorch.org/whl/cu130"
explicit = true

[project.scripts]
inject-ids = "gelato.dataset.inject_mxl_ids:main"
render-dataset = "gelato.dataset.render_dataset:main"
paginate-mxl = "gelato.dataset.paginate_mxl:main"
mxl2abc = "gelato.dataset.mxl2abc:main"
clean-dataset = "gelato.dataset.clean_dataset:main"
strip-abc = "gelato.dataset.strip_abc:main"
easy-writer = "gelato.dataset.easy_writer:main"
inter-writer = "gelato.dataset.inter_writer:main"
pretrain-gelato = "gelato.train.pretrain:main"
train-gelato = "gelato.train.train:main"
eval-gelato = "gelato.train.eval:main"
